WIP: Fix remaining -Wformat-signedness warnings

+++ contrib/isn/isn.c
@@ -890,7 +890,7 @@ eanbadcheck:
 			ereport(ERROR,
 					(errcode(ERRCODE_INVALID_TEXT_REPRESENTATION),
 					 errmsg("invalid check digit for %s number: \"%s\", should be %c",
-							isn_names[accept], str, (rcheck == 10) ? ('X') : (rcheck + '0'))));
+							isn_names[accept], str, (rcheck == 10) ? ('X') : (int) (rcheck + '0'))));
 		}
 	}
 	return false;
